+++ framework/Database/Migrations/2023_12_29_085321_create_migrations_table.php
@@ -7,7 +7,7 @@ return new class extends Migration
 {
     public function run(): void
     {
-        Database::query(
+        Database::unprepared(
             'CREATE TABLE migrations (' .
             'id INT auto_increment,' .
             'name VARCHAR(255) NOT NULL,' .