+++ framework/Database/Migrations/2023_12_29_100125_create_users_table.php
@@ -7,7 +7,7 @@ return new class extends Migration
 {
     public function run(): void
     {
-        Database::query(
+        Database::unprepared(
             'CREATE TABLE users (' .
             'id INT auto_increment,' .
             'firstname VARCHAR(30) NOT NULL,' .