+++ framework/Database/Migrations/2024_01_01_013858_create_user_roles_table.php
@@ -7,7 +7,7 @@ return new class extends Migration
 {
     public function run(): void
     {
-        Database::query(
+        Database::unprepared(
             'CREATE TABLE userRoles (' .
             'id INT auto_increment,' .
             'userId INT NOT NULL,' .